Change apdu payload length to proper size (8 bits)

+++ src_features/signTx/cmd_signTx.c
@@ -7,7 +7,7 @@
 void handleSign(uint8_t p1,
                 uint8_t p2,
                 const uint8_t *workBuffer,
-                uint16_t dataLength,
+                uint8_t dataLength,
                 unsigned int *flags,
                 unsigned int *tx) {
     UNUSED(tx);
