+++ src_features/stark_sign/cmd_stark_sign.c
@@ -15,7 +15,7 @@
 void handleStarkwareSignMessage(uint8_t p1,
                                 uint8_t p2,
                                 uint8_t *dataBuffer,
-                                uint16_t dataLength,
+                                uint8_t dataLength,
                                 unsigned int *flags,
                                 __attribute__((unused)) unsigned int *tx) {
     uint8_t privateKeyData[INT256_LENGTH];
