+++ src_features/stark_unsafe_sign/cmd_stark_unsafe_sign.c
@@ -9,7 +9,7 @@
 void handleStarkwareUnsafeSign(uint8_t p1,
                                uint8_t p2,
                                const uint8_t *dataBuffer,
-                               uint16_t dataLength,
+                               uint8_t dataLength,
                                unsigned int *flags,
                                __attribute__((unused)) unsigned int *tx) {
     uint8_t privateKeyData[INT256_LENGTH];
